requirements.txt
langchain
langchain-openai
langsmith
python-dotenv
yfinance
pandas
pandas-ta
ddgs
transformers
torch
textblob
numpy==1.26.4
scipy